# FIX: improve performance of UserStat.ensure_consistency (#21044)